+++ server/modules/textobjects/src/word.rs
@@ -324,7 +324,7 @@ mod tests {
             &self,
             _cmd: &CommandId,
             _ctx: &CommandContext,
-            _kernel: &mut KernelContext,
+            _kernel: &KernelContext,
         ) -> Option<CommandResult> {
             Some(CommandResult::Success)
         }