+++ server/modules/vim/src/operators/change.rs
@@ -200,7 +200,7 @@ mod tests {
                 &self,
                 _: &CommandId,
                 _: &CommandContext,
-                _: &mut KernelContext,
+                _: &KernelContext,
             ) -> Option<CommandResult> {
                 Some(CommandResult::Success)
             }