+++ server/modules/vim/src/operators/delete.rs
@@ -275,7 +275,7 @@ mod tests {
                 &self,
                 _: &CommandId,
                 _: &CommandContext,
-                _: &mut KernelContext,
+                _: &KernelContext,
             ) -> Option<CommandResult> {
                 Some(CommandResult::Success)
             }