+++ server/modules/vim/src/operators/yank.rs
@@ -135,7 +135,7 @@ mod tests {
                 &self,
                 _: &CommandId,
                 _: &CommandContext,
-                _: &mut KernelContext,
+                _: &KernelContext,
             ) -> Option<CommandResult> {
                 Some(CommandResult::Success)
             }